requirements.txt
# Core dependencies
requests>=2.25.0
cryptography>=3.4.0
urllib3>=1.26.0

# Optional dependencies for development and testing
python-dotenv>=0.19.0
# pytest>=6.0
# pytest-cov>=2.0
# black>=21.0
# flake8>=3.8
# mypy>=0.800
